Move test_definition to test_codeset
